Move prebuilt binaries to GitHub releases
This changes the workflow so that Travis CI pushes a new tag for every
commit to the repository and then for that commit it will package the
binaries and upload them to that commit for GitHub.

The GitHub API token was generated by the SirKnightly account which
should be able to create the releases.